pyproject.toml
[tool.poetry]
name = "squeaky-hinge"
version = "0.1.0"
description = "Reliable notifications for Hinge messages"
authors = ["Radian LLC <contact+squeaky-hinge@radian.codes>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
websocket-client = "^1.3.3"
Flask = "^2.2.2"
humanize = "^4.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
